refactor: 구형 컨트롤러 패키지 분리
