+++ app/src/main/java/com/techcourse/controller/asis/RegisterController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.asis;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
@@ -10,7 +10,7 @@ public class RegisterController implements Controller {
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        final User user = new User(2,
+        final User user = new User(
                 req.getParameter("account"),
                 req.getParameter("password"),
                 req.getParameter("email"));
