+++ app/src/main/java/com/techcourse/controller/asis/RegisterViewController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
